requirements.txt
python-docx
fastapi[standard]
langchain_qdrant
langchain_core
langchain_google_genai
langchain_groq
langchain_huggingface
langchain_text_splitters
langgraph
numpy
pydantic
pydantic[email]
PyPDF2
python-dotenv
bcrypt
python_jose
redis
SQLAlchemy
starlette
alembic
cryptography 
pymysql
sentence-transformers
python-multipart